Feature for issue #2 [Empty Cell이 하나도 없는 경우GameOver]
