+++ drawBoard.c
@@ -52,10 +52,6 @@ void updateBoardCells(struct Square m[][SIZE], int value, int used) {
 
 void initBoard(struct Square m[][SIZE]) {
 	updateBoardCells(m, 0, 0);
-	// int i, randomSquares = 3;
-	// for (i = 0; i < randomSquares; i++) {
-	// 	addRandomSquare(m);
-	// }
 }
 void cleanCells(struct Square m[][SIZE]) {
 	updateBoardCells(m, -1, 0);
